change code for read

+++ JackRESTfulCommunication/JackRESTArtist.m
@@ -13,147 +13,149 @@
 
 - (id)init {
     if ((self = [super init])) {
-        // Initialization code here.
     }
-    
+
     return self;
 }
 
 - (void)dealloc {
-    // Clean-up code here.
-    
+
     [super dealloc];
 }
 
 @synthesize name;
 @synthesize ip;
 @synthesize roomName;
-//@synthesize index;
 @synthesize status;
 @synthesize message;
 
--(id) init:(NSXMLNode *) node{
+- (id)init:(NSXMLNode *)node {
     if ((self = [super init])) {
-        NSError * err = nil;
-        NSXMLNode * element = nil;
-        
-        if([[node name] isEqualToString:@"message"]){
+        NSError *err = nil;
+        NSXMLNode *element = nil;
+
+        if ([[node name] isEqualToString:@"message"]) {
             message = [[[JackRESTMessage alloc] init:node] autorelease];
         }
-        else{
+        else {
             element = [[node nodesForXPath:@"Name" error:&err] objectAtIndex:0];
             name = [element stringValue];
             element = [[node nodesForXPath:@"IP" error:&err] objectAtIndex:0];
             ip = [element stringValue];
             element = [[node nodesForXPath:@"RoomName" error:&err] objectAtIndex:0];
             roomName = [element stringValue];
-            //element = [[node nodesForXPath:@"Index" error:&err] objectAtIndex:0];
-            //index = [element stringValue];
             element = [[node nodesForXPath:@"Status" error:&err] objectAtIndex:0];
             status = [[element stringValue] integerValue];
         }
-        
+
         [element release];
         [err release];
     }
-    
+
     return self;
 }
--(id) init:(NSString *) artistName 
-        ip:(NSString *) artistIP 
-  roomName:(NSString *) artistRoomName{
+
+- (id)init:(NSString *)artistName
+        ip:(NSString *)artistIP
+  roomName:(NSString *)artistRoomName {
     if ((self = [self init])) {
         name = artistName;
         ip = artistIP;
         roomName = artistRoomName;
     }
-    
+
     return self;
 }
--(NSString *) toXmlString{
-    return [NSString stringWithFormat:@"<?xml version='1.0' encoding='UTF-8' ?><artist><Name>%@</Name><IP>%@</IP><RoomName>%@</RoomName><Status>%d</Status></artist>",name,ip,roomName,status];
+
+- (NSString *)toXmlString {
+    return [NSString stringWithFormat:@"<?xml version='1.0' encoding='UTF-8' ?><artist><Name>%@</Name><IP>%@</IP><RoomName>%@</RoomName><Status>%d</Status></artist>", name, ip, roomName, (int) status];
 }
 
-+(NSString *) getCreateArtistUrl{
-    return [[NSString stringWithFormat:@"%@/artists.xml",[JackRESTRoom getHostUrl]] autorelease];
++ (NSString *)getCreateArtistUrl {
+    return [[NSString stringWithFormat:@"%@/artists.xml", [JackRESTRoom getHostUrl]] autorelease];
 }
-+(NSString *) getArtistUrl:(NSString *) artistName{
-    return [[NSString stringWithFormat:@"%@/artists/%@.xml",[JackRESTRoom getHostUrl],artistName] autorelease];
+
++ (NSString *)getArtistUrl:(NSString *)artistName {
+    return [[NSString stringWithFormat:@"%@/artists/%@.xml", [JackRESTRoom getHostUrl], artistName] autorelease];
 }
-+(NSString *) getUpdateArtistStatusUrl:(NSString *) artistName
-                                status:(NSInteger) artistStatus{
-    return [[NSString stringWithFormat:@"%@/artists/%@.%d.xml",[JackRESTRoom getHostUrl],artistName,artistStatus] autorelease];
+
++ (NSString *)getUpdateArtistStatusUrl:(NSString *)artistName
+                                status:(NSInteger)artistStatus {
+    return [[NSString stringWithFormat:@"%@/artists/%@.%d.xml", [JackRESTRoom getHostUrl], artistName, (int) artistStatus] autorelease];
 }
-+(id) createArtist:(JackRESTArtist *) artist{
-    NSString * data = [HttpRequestHelper sendPOSTRequestWithXmlText:[JackRESTArtist getCreateArtistUrl] 
-                                                            xmlText:[artist toXmlString]];
-    if(data == nil){
-        JackRESTArtist * artistEmptyInfo = [[JackRESTArtist alloc] init];
+
++ (id)createArtist:(JackRESTArtist *)artist {
+    NSString *data = [HttpRequestHelper sendPOSTRequestWithXmlText:[JackRESTArtist getCreateArtistUrl]
+                                                           xmlText:[artist toXmlString]];
+    if (data == nil) {
+        JackRESTArtist *artistEmptyInfo = [[JackRESTArtist alloc] init];
         [data release];
         return [artistEmptyInfo autorelease];
     }
-    
-    NSError * err = nil;
-    NSXMLDocument * xmlDocument = [[NSXMLDocument alloc] initWithXMLString:data 
-                                                                   options:(NSXMLNodePreserveWhitespace | NSXMLNodePreserveCDATA) 
-                                                                     error:&err];
-	if(err != nil){
-        JackRESTArtist * artistEmptyInfo = [[JackRESTArtist alloc] init];
-		[err release];
-		[xmlDocument release];
-		return [artistEmptyInfo autorelease];
-	}
-    
-    JackRESTArtist * artistInfo = [[JackRESTArtist alloc] init:[xmlDocument rootElement]];
+
+    NSError *err = nil;
+    NSXMLDocument *xmlDocument = [[NSXMLDocument alloc] initWithXMLString:data
+                                                                  options:(NSXMLNodePreserveWhitespace | NSXMLNodePreserveCDATA)
+                                                                    error:&err];
+    if (err != nil) {
+        JackRESTArtist *artistEmptyInfo = [[JackRESTArtist alloc] init];
+        [err release];
+        [xmlDocument release];
+        return [artistEmptyInfo autorelease];
+    }
+
+    JackRESTArtist *artistInfo = [[JackRESTArtist alloc] init:[xmlDocument rootElement]];
     [err release];
     [xmlDocument release];
-	return [artistInfo autorelease];
+    return [artistInfo autorelease];
 }
-+(id) getArtist:(NSString *) artistName{
-    NSString * data = [HttpRequestHelper sendGETRequest:[JackRESTArtist getArtistUrl:artistName]];
-    if(data == nil){
-        JackRESTArtist * artistEmptyInfo = [[JackRESTArtist alloc] init];
+
++ (id)getArtist:(NSString *)artistName {
+    NSString *data = [HttpRequestHelper sendGETRequest:[JackRESTArtist getArtistUrl:artistName]];
+    if (data == nil) {
+        JackRESTArtist *artistEmptyInfo = [[JackRESTArtist alloc] init];
         [data release];
         return [artistEmptyInfo autorelease];
     }
-    
-    NSError * err = nil;
-	NSXMLDocument * xmlDocument = [[NSXMLDocument alloc] initWithXMLString: data
-                                                                   options:(NSXMLNodePreserveWhitespace | NSXMLNodePreserveCDATA) 
-                                                                     error:&err];
-	if(err != nil){
-		JackRESTArtist * artistEmptyInfo = [[JackRESTArtist alloc] init];
+
+    NSError *err = nil;
+    NSXMLDocument *xmlDocument = [[NSXMLDocument alloc] initWithXMLString:data
+                                                                  options:(NSXMLNodePreserveWhitespace | NSXMLNodePreserveCDATA)
+                                                                    error:&err];
+    if (err != nil) {
+        JackRESTArtist *artistEmptyInfo = [[JackRESTArtist alloc] init];
         [err release];
-		[xmlDocument release];
-		return [artistEmptyInfo autorelease];
-	}
-    
-    JackRESTArtist * artistInfo = [[JackRESTArtist alloc] init:[xmlDocument rootElement]];
+        [xmlDocument release];
+        return [artistEmptyInfo autorelease];
+    }
+
+    JackRESTArtist *artistInfo = [[JackRESTArtist alloc] init:[xmlDocument rootElement]];
     [err release];
     [xmlDocument release];
-	return [artistInfo autorelease];
+    return [artistInfo autorelease];
 }
-+(id) updateArtistStatus:(NSString *) artistName 
-                  status:(NSInteger) artistStatus{
-    NSString * data = [HttpRequestHelper sendGETRequest:[JackRESTArtist getUpdateArtistStatusUrl:artistName status:artistStatus]];
-    if(data == nil){
-        JackRESTArtist * artistEmptyInfo = [[JackRESTArtist alloc] init];
+
++ (id)updateArtistStatus:(NSString *)artistName
+                  status:(NSInteger)artistStatus {
+    NSString *data = [HttpRequestHelper sendGETRequest:[JackRESTArtist getUpdateArtistStatusUrl:artistName status:artistStatus]];
+    if (data == nil) {
+        JackRESTArtist *artistEmptyInfo = [[JackRESTArtist alloc] init];
         [data release];
         return [artistEmptyInfo autorelease];
     }
-    
-    NSError * err = nil;
-	NSXMLDocument * xmlDocument = [[NSXMLDocument alloc] initWithXMLString:data 
-                                                                   options:(NSXMLNodePreserveWhitespace | NSXMLNodePreserveCDATA) 
-                                                                     error:&err];
-	if(err != nil){
-		JackRESTArtist * artistEmptyInfo = [[JackRESTArtist alloc] init];
+
+    NSError *err = nil;
+    NSXMLDocument *xmlDocument = [[NSXMLDocument alloc] initWithXMLString:data
+                                                                  options:(NSXMLNodePreserveWhitespace | NSXMLNodePreserveCDATA)
+                                                                    error:&err];
+    if (err != nil) {
+        JackRESTArtist *artistEmptyInfo = [[JackRESTArtist alloc] init];
         [err release];
-		[xmlDocument release];
-		return [artistEmptyInfo autorelease];
-	}
-    
-    JackRESTArtist * artistInfo = [[JackRESTArtist alloc] init:[xmlDocument rootElement]];
+        [xmlDocument release];
+        return [artistEmptyInfo autorelease];
+    }
+
+    JackRESTArtist *artistInfo = [[JackRESTArtist alloc] init:[xmlDocument rootElement]];
     [err release];
     [xmlDocument release];
     return [artistInfo autorelease];
